add activity feed to user profile, close #76
